Migrate to contrastsecurity yamlbeans 1.17
The esotericsoftware yamlbeans version seems to be semi-abandoned, or team have lost ability to publish to Maven based on
https://github.com/EsotericSoftware/yamlbeans/pull/164#issuecomment-1702558881 There are a couple of CVEs mitigated here,
but we need to re-enable anchors/aliases as these are core functionality.

Also minor tweaks needed to tests due to yamlbeans correcting the support for block scalars in 1.16. See https://yaml-multiline.info/